Move logic to own file
